fix FlashSaleItemResponse

+++ Backend/src/main/java/com/Hyperfume/Backend/dto/response/FlashSaleItemResponse.java
@@ -21,7 +21,7 @@ public class FlashSaleItemResponse {
 
     BigDecimal originalMinPrice;
 
-    double discountedPrice;
+    BigDecimal discountedPrice;
 
     int quantityLimit;
